+++ samplesheets/vueapp/tests/unit/HeaderEditRenderer.spec.js
@@ -183,17 +183,6 @@ describe('HeaderEditRenderer.vue', () => {
     expect(header.find('.sodar-ss-col-config-btn').exists()).toBe(true)
   })
 
-  it('renders header when user is not allowed to edit column', async () => {
-    const sc = copy(sodarContext)
-    sc.perms.edit_config = false
-    const wrapper = mountSheetTable({ sodarContext: sc })
-    await waitAG(wrapper)
-    await waitRAF()
-
-    const header = wrapper.find('.sodar-ss-header-edit')
-    expect(header.find('.sodar-ss-col-config-btn').exists()).toBe(false)
-  })
-
   // TODO: Test data provided to modalComponent per column type
   // TODO: How to access/modify renderer data?
 })